+++ src/cli/contract/activate/index.js
@@ -2,11 +2,11 @@ const shell = require('shelljs');
 const chalk = require('chalk');
 const { cli, action } = require('core/cli');
 const { getTopRunningChains } = require('core/forge-process');
-const { execute, run } = require('./deactivate');
+const { execute, run } = require('./activate');
 
 cli(
-  'protocol:deactivate [name|adderss]',
-  'Deactivate a transaction protocol',
+  'contract:activate [name|address]',
+  'Activate a contract by name or address',
   input => action(execute, run, input),
   {
     requirements: {
@@ -22,10 +22,10 @@ cli(
       '--help': () => {
         shell.echo(`
 Examples:
-  - ${chalk.cyan('forge deactivate transfer')}         deactivate protocol by name
+  - ${chalk.cyan('forge activate transfer')}         Activate contract by name
   - ${chalk.cyan(
-    'forge deactivate z2E3vuTEWpE7PhmVP61eZWJXutgxMHeimwqtv'
-  )}   deactivate protocol by address
+    'forge activate z2E3vuTEWpE7PhmVP61eZWJXutgxMHeimwqtv'
+  )}   Activate contract by address
         `);
       },
     },