+++ src/cli/contract/index.js
@@ -1,5 +1,6 @@
 exports.activate = require('./activate');
 exports.compile = require('./compile');
+exports.create = require('./create');
 exports.deactivate = require('./deactivate');
 exports.deploy = require('./deploy');
 exports.list = require('./list');