+++ src/cli/contract/list/index.js
@@ -3,7 +3,7 @@ const { cli, action } = require('core/cli');
 const { getTopRunningChains } = require('core/forge-process');
 const { execute, run } = require('./list');
 
-cli('protocol:ls', 'List transaction protocols', input => action(execute, run, input), {
+cli('contract:ls', 'List contracts', input => action(execute, run, input), {
   requirements: {
     forgeRelease: true,
     runningNode: true,